requirements.txt
flask
bs4
ultralytics
pillow
gunicorn